backend/requirements.txt
pyvisa
pyvisa-py
pandas
numpy
pure-python-adb
pyyaml
pytest
fastapi
uvicorn[standard]
websocketspython-multipart
